base_requirement.txt
Django
djangorestframework
flake8
Pillow
psycopg2
django-cors-headers
python-dotenv